etl/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
xmltodict==0.13.0
click==8.1.7